% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translateToBraille}
\alias{translateToBraille}
\title{Translate text to Braille}
\usage{
translateToBraille(text, table)
}
\arguments{
\item{text}{A character string to translate.}

\item{table}{The Braille table to use (e.g., "en-ueb-g2.ctb").}
}
\value{
A character string of translated Braille.
}
\description{
Translate text to Braille
}
