% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaleSports.R
\docType{data}
\name{yaleSports}
\alias{yaleSports}
\title{Yale Sports Dataset}
\format{
A data frame with 1,228 rows and 6 variables:
\describe{
  \item{Year}{Year of sport team operation}
  \item{Sport}{Student athlete team membership}
  \item{ScramName}{Student athlete full name alpha-sorted and lower-cased}
  \item{Class}{Student athlete class standing}
  \item{Height}{Student athlete height in inch units}
  \item{Weight}{Student athlete weight in pounds units}
}
}
\source{
https://yalebulldogs.com (Retrieved Summer, 2018)
}
\usage{
yaleSports
}
\description{
Adapted from classic dataset from Yale School of Public Health Introduction to Biostatistics courses (EPH-505, EMPH-535, PUBH-505). Date retrieved from Yale Athletics pages.
}
\keyword{datasets}
