% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_br.R
\name{barplot_br}
\alias{barplot_br}
\title{Braile on Bar Plot}
\usage{
barplot_br(...)
}
\arguments{
\item{...}{plot parameters as would typically be used in barplot()}
}
\value{
two-page .pdf file
}
\description{
Create Braille-enabled Bar Plot
}
\examples{
\dontrun{
yale = data.frame(yaleSports)
tbl = table(yale$Sport)
barplot_br(tbl,
           over="Bar Chart",
           main="Yale Athletics Dataset",
           xlab="x-axis = Sport",
           ylab="y-axis = Count (n athletes)",
           stem="newBar")
}
\dontrun{
air = airquality
air$Feel = cut(air$Temp,breaks=c(0,65,80,1000),labels=c("cool","warm","hot"))
tbl = table(air$Feel)
barplot_br(tbl,
           over="Bar Chart",
           main="Air Quality: NYC (1973)",
           xlab="x-axis = Temperature Feel",
           ylab="y-axis = Count (n days)",
           stem="airBar",
           cex.main=1.5,
           cex.axis=3.5,
           col="purple")
}
}
