% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_br.R
\name{boxplot_br}
\alias{boxplot_br}
\title{Braile on Box Plot}
\usage{
boxplot_br(...)
}
\arguments{
\item{...}{plot parameters as would typically be used in boxplot()}
}
\value{
two-page .pdf file
}
\description{
Create Braille-enabled Box Plot
}
\examples{
\dontrun{
yale = data.frame(yaleSports)
boxplot_br(Height~Sport,data=yale,
      over="Box Plot",
      main="Yale Athletics Dataset",
      xlab="x-axis = Sport",
      ylab="y-axis = Height (inches)",
      stem="newBox")
}
\dontrun{
air = airquality
air$Month = as.factor(air$Month)
boxplot_br(Ozone~Month,data=air,
      lwd=2,pch=16,las=3,lty=1,
      over="Box Plot",
      main="Air Quality: NYC (1973)",
      xlab="x-axis = Month (numeric)",
      ylab="y-axis = Ozone (ppb)",
      stem="airBox",
      cex.main=1.5,
      cex.axis=1.5,
      col="chocolate1")
}
}
