% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_br.R
\name{hist_br}
\alias{hist_br}
\title{Braile on Histogram}
\usage{
hist_br(x, breaks = 20, ...)
}
\arguments{
\item{x}{numerical data for summarizing in histogram}

\item{breaks}{number (plus one) of bins in the histogram}

\item{...}{plot parameters as would typically be used in hist()}
}
\value{
two-page .pdf file
}
\description{
Create Braille-enabled Histogram
}
\examples{
\dontrun{
yale = data.frame(yaleSports)
hist_br(yale$Weight,
        over="Histogram",
        main="Yale Athletics Dataset",
        xlab="x-axis = Bodyweight (pounds)",
        ylab="y-axis = Count (n athletes)",
        stem="newHist")
}
\dontrun{
air=airquality
hist_br(air$Temp,
       over="Histogram",
       main="Air Quality: NYC (1973)",
       xlab="x-axis = Temperature (deg)",
       ylab="y-axis = Count (n days)",
       stem="airHist",
       cex.main=2,
       cex.axis=2,
       col="deeppink1")
}
}
