% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_br.R
\name{pie_br}
\alias{pie_br}
\title{Braile on Pie Chart}
\usage{
pie_br(...)
}
\arguments{
\item{...}{plot parameters as would typically be used in pie()}
}
\value{
two-page .pdf file
}
\description{
Create Braille-enabled Pie Chart
}
\examples{
\dontrun{
yale = data.frame(yaleSports)
tbl = table(yale$Sport)
pie_br(tbl,
       over="Pie Chart",
       main="Yale Athletics Dataset",
       xlab=" ",
       ylab=" ",
       stem="newPie")
}
\dontrun{
air = airquality
air$Feel = cut(air$Temp,breaks=c(0,65,80,1000),labels=c("cool","warm","hot"))
tbl = table(air$Feel)
pie_br(tbl,
       over="Pie Chart",
       main="Air Quality: NYC (1973)",
       xlab="",
       ylab="",
       stem="airPie",
       cex.main=2,
       cex=1)
}
}
