% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_br.R
\name{plot_br}
\alias{plot_br}
\title{Braile on Scatter or Line Plots}
\usage{
plot_br(x, y = NULL, data = NULL, ...)
}
\arguments{
\item{x}{variable for plotting (independent variable, if y is not NULL)}

\item{y}{dependent variable for plotting}

\item{data}{data-frame in case formula argument utilized}

\item{...}{plot parameters as would typically be used in plot()}
}
\value{
two-page .pdf file
}
\description{
Create Braille-enabled Scatter or Line Plot
}
\examples{
\dontrun{
yale = data.frame(yaleSports)
plot_br(Weight~Height,data=yale,
        over="Scatter Plot",
        main="Yale Athletics Dataset",
        xlab="x-axis = Height (inches)",
        ylab="y-axis = Weight (pounds)",
        stem="plotScatter")
}
\dontrun{
yale = data(yaleSports)
plot_br(sort(yale$Weight[1:20]),
        over="Scatter Plot",
        main="Yale Athletics Dataset",
        xlab="Athlete (sorted)",
        ylab="y-axis = Weight (pounds)",
        type="b",
        stem="plotLine")
}
\dontrun{
air = airquality
plot_br(air$Temp,air$Ozone,
        over="Scatter Plot",
        main="Air Quality: NYC (1973)",
        xlab="x-axis = Temperature (deg)",
        ylab="y-axis = Ozone (ppb)",
        stem="airPlot",
        cex.main=1.5,
        cex.axis=1.5,
        col="red",
        pch=16)
}
}
